Move SlotMetadata from models.py to websocket.py
SlotMetadata is the per-slot shape inside websocket subscription
responses (typed/JSON-serializable values parsed from entity states).
Used only by websocket.py — moved next to its consumer.

models.py now contains only genuinely cross-cutting types:
SlotCode (enum used by sync, providers, coordinator, data, websocket,
tests), LockCodeManagerConfigEntryData (runtime data), the
LockCodeManagerConfigEntry type alias, and SlotEntities (used by
websocket only today, but a shape that other modules might naturally
want — kept in models.py for now since it's a generic slot-entity
container).

The "canonical home for cross-cutting types" claim in models.py is
now truthful instead of aspirational.

602 tests pass.

Entire-Checkpoint: 39ec9353b556
